split net operator and serial
